criando nova tarefa de RTD e adcionando mais sugestões do PR
