+++ Core/Inc/RTD.h
@@ -1,0 +1,16 @@
+/*
+ * RTD_handler.h
+ *
+ *  Created on: Aug 15, 2021
+ *      Author: Felipe Telles
+ */
+
+#ifndef INC_RTD_H_
+#define INC_RTD_H_
+
+#define tempo_sirene 1000
+
+void exit_RTD();
+void aciona_sirene();
+
+#endif /* INC_RTD_H_ */
