+++ Core/Inc/main_task.h
@@ -8,13 +8,7 @@
 #ifndef INC_MAIN_TASK_H_
 #define INC_MAIN_TASK_H_
 
-#include "global_instances.h"
-#include "global_definitions.h"
-#include "cmsis_os.h"
 
-#define tempo_sirene 1000
 
-void exit_RTD();
-void aciona_sirene();
 
 #endif /* INC_MAIN_TASK_H_ */
