+++ Core/Src/external_interrupt_handler.c
@@ -23,7 +23,7 @@ void HAL_GPIO_EXTI_Callback(uint16_t GPIO_Pin){
 			break;
 
 		case B_RTD_Pin:
-			osThreadFlagsSet(t_main_taskHandle, RTD_BTN_PRESSED_FLAG);
+			osThreadFlagsSet(t_RTD_handlerHandle, RTD_BTN_PRESSED_FLAG);
 			break;
 
 		case B_MODO_Pin:
