+++ ECU.ioc
@@ -86,7 +86,7 @@ FREERTOS.FootprintOK=true
 FREERTOS.IPParameters=Tasks01,configMINIMAL_STACK_SIZE,FootprintOK,Queues01,configMAX_TASK_NAME_LEN,Mutexes01,configENABLE_FPU,configTOTAL_HEAP_SIZE,configUSE_MALLOC_FAILED_HOOK,configUSE_STATS_FORMATTING_FUNCTIONS,configCHECK_FOR_STACK_OVERFLOW,BinarySemaphores01
 FREERTOS.Mutexes01=m_state_parameter_mutex,Dynamic,NULL
 FREERTOS.Queues01=q_speed_message,16,speed_message_t,0,Dynamic,NULL,NULL;q_torque_message,16,torque_message_t,0,Dynamic,NULL,NULL;q_ref_torque_message,16,ref_torque_t,0,Dynamic,NULL,NULL;q_datalog_message,128,datalog_message_t,0,Dynamic,NULL,NULL;q_debugleds_message,16,debugled_message_t,0,Dynamic,NULL,NULL
-FREERTOS.Tasks01=t_main_task,24,1024,main_task,As weak,NULL,Dynamic,NULL,NULL;t_controle,8,1024,controle,As external,NULL,Dynamic,NULL,NULL;t_datalogger,8,1024,datalogger,As external,NULL,Dynamic,NULL,NULL;t_throttle_read,8,1024,throttle_read,As external,NULL,Dynamic,NULL,NULL;t_steering_read,8,1024,steering_read,As external,NULL,Dynamic,NULL,NULL;t_speed_calc,8,1024,speed_calc,As external,NULL,Dynamic,NULL,NULL;t_odometer_calc,8,1024,odometer_calc,As external,NULL,Dynamic,NULL,NULL;t_thr_handler,8,1024,throttle_handler,As external,NULL,Dynamic,NULL,NULL;t_torque_manager,8,1024,torque_manager,As external,NULL,Dynamic,NULL,NULL;t_debugleds,8,1024,debugleds,As external,NULL,Dynamic,NULL,NULL;t_rgb_led,8,1024,rgb_led,As external,NULL,Dynamic,NULL,NULL;t_seleciona_modo,8,1024,seleciona_modo,As external,NULL,Dynamic,NULL,NULL
+FREERTOS.Tasks01=t_main_task,24,1024,main_task,As weak,NULL,Dynamic,NULL,NULL;t_controle,8,1024,controle,As external,NULL,Dynamic,NULL,NULL;t_datalogger,8,1024,datalogger,As external,NULL,Dynamic,NULL,NULL;t_throttle_read,8,1024,throttle_read,As external,NULL,Dynamic,NULL,NULL;t_steering_read,8,1024,steering_read,As external,NULL,Dynamic,NULL,NULL;t_speed_calc,8,1024,speed_calc,As external,NULL,Dynamic,NULL,NULL;t_odometer_calc,8,1024,odometer_calc,As external,NULL,Dynamic,NULL,NULL;t_thr_handler,8,1024,throttle_handler,As external,NULL,Dynamic,NULL,NULL;t_torque_manager,8,1024,torque_manager,As external,NULL,Dynamic,NULL,NULL;t_debugleds,8,1024,debugleds,As external,NULL,Dynamic,NULL,NULL;t_rgb_led,8,1024,rgb_led,As external,NULL,Dynamic,NULL,NULL;t_seleciona_modo,8,1024,seleciona_modo,As external,NULL,Dynamic,NULL,NULL;t_RTD_handler,8,1024,RTD_handler,As external,NULL,Dynamic,NULL,NULL
 FREERTOS.configCHECK_FOR_STACK_OVERFLOW=2
 FREERTOS.configENABLE_FPU=1
 FREERTOS.configMAX_TASK_NAME_LEN=32
